refactor: Rename to discount for consistent naming
