+++ app/Data/Discounts/DiscountData.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace App\Data\Promotions;
+namespace App\Data\Discounts;
 
 use Carbon\Carbon;
 use Spatie\LaravelData\Attributes\DataCollectionOf;
 use Spatie\LaravelData\Data;
 
-class PromotionData extends Data
+class DiscountData extends Data
 {
     public function __construct(
         public Carbon $start,
         public Carbon $end,
-        #[DataCollectionOf(PromotionTierData::class)] public array $tiers
+        #[DataCollectionOf(DiscountTierData::class)] public array $tiers
     ) {}
 }
